Move platform binary pruning to install:bundle script phase
Replace the Vite closeBundle plugin with a standalone script that runs
during install:bundle, before Vite copies node_modules to dist. This
ensures foreign-platform binaries are removed from source node_modules
and never make it into the build output. Follows the same pattern as
the existing remove-fs-ext-other-platform-binaries.mjs script.
