+++ apps/cli/package.json
@@ -31,7 +31,7 @@
 	},
 	"scripts": {
 		"build": "vite build --config ./vite.config.ts",
-		"install:bundle": "npm install --omit=dev --no-package-lock --no-progress --install-links --no-workspaces && patch-package && node ../../scripts/remove-fs-ext-other-platform-binaries.mjs",
+		"install:bundle": "npm install --omit=dev --no-package-lock --no-progress --install-links --no-workspaces && patch-package && node ../../scripts/remove-fs-ext-other-platform-binaries.mjs && node ../../scripts/remove-other-platform-binaries.mjs",
 		"package": "npm run install:bundle && npm run build",
 		"watch": "vite build --config ./vite.config.ts --watch"
 	},
